translation auto-check started
