+++ wordengine/admin.py
@@ -32,3 +32,4 @@ admin.site.register(models.LexemeParameter)
 admin.site.register(models.SyntCatsInLanguage)
 admin.site.register(models.Theme)
 admin.site.register(models.Dictionary)
+admin.site.register(models.WSInDict)
