+++ wordengine/templates/wordengine/base.html
@@ -12,7 +12,8 @@
 <!--    <a href="">Quickly add translations</a> | -->
 <!--    <a href="{% url 'wordengine:add_wordform_lexeme' %}">Add word (directly)</a> | -->
     <a href="{% url 'wordengine:list_lexeme_entry' %}">Lexeme entries</a> |
-    <a href="{% url 'wordengine:admin' %}">Administer</a></div>
+    <a href="{% url 'wordengine:admin' %}">Administer</a> |
+    <a href="{% url 'wordengine:admin' %}/login/">Admin login</a></div>
     <!--<a href="{% url 'admin:index' %}">Django admin</a></div>-->
 <hr />
